fix: added changes

+++ client/src/components/Message/Message.js
@@ -8,16 +8,24 @@ import { messageFormSchema } from './validation';
 import './styles.css';
 import MessageForm from '../MessageForm/MessageForm';
 import Solutionlist from '../Solutions/Solutionlist';
+import { Badge, Button, Checkbox, Input, Modal, Radio, Row, Spacer, Text, Textarea } from '@nextui-org/react';
+import { Categories } from '../../constants';
+import ReactMarkdown from 'react-markdown'
 
 
 const Message = ({ message, auth, deleteMessage, editMessage, clearMessageError }) => {
   const [isEdit, setIsEdit] = useState(false);
   const [openModal, setOpenModal] = useState(false);
   const [openSolution, setOpenSolution] = useState(false);
-
+  const [page, setPage] = useState(1);
+  const [selected, setSelected] = useState([]);
+  const [selectedDifficulty, setSelectedDifficulty] = useState('');
+  const [selectPublic, setselectedPublic] = useState(true);
 
   const handleDelete = (e, id) => {
     e.preventDefault();
+    // eslint-disable-next-line no-restricted-globals
+    if (!window.confirm("Are you sure you want to delete?")) return;
     if (!isEdit) {
       deleteMessage(id);
     }
@@ -25,13 +33,18 @@ const Message = ({ message, auth, deleteMessage, editMessage, clearMessageError 
 
   const handleClickEdit = (e) => {
     e.preventDefault();
+    formik.setFieldValue('title', message.title);
     formik.setFieldValue('text', message.text);
     formik.setFieldValue('incentive', message.incentive);
-    formik.setFieldValue('category', message.category);
+    setSelected(message.category);
+    setSelectedDifficulty(message.difficulty);
+    setselectedPublic(message.public);
     setIsEdit((oldIsEdit) => !oldIsEdit);
   };
 
   const handleClose = () => setOpenModal(false)
+  const closeHandler = () => setIsEdit(false);
+
 
   const formik = useFormik({
     enableReinitialize: true,
@@ -43,8 +56,9 @@ const Message = ({ message, auth, deleteMessage, editMessage, clearMessageError 
     },
     validationSchema: messageFormSchema,
     onSubmit: (values, { resetForm }) => {
-      editMessage(values.id, { text: values.text, incentive: values.incentive, category: values.category });
-      setIsEdit(false);
+      editMessage(values.id, { title: values.title, text: values.text, incentive: values.incentive, category: selected, public: selectPublic, difficulty: selectedDifficulty });
+      closeHandler();
+      setPage(1);
       // resetForm();
     },
   });
@@ -62,122 +76,218 @@ const Message = ({ message, auth, deleteMessage, editMessage, clearMessageError 
   return (
     <div className={message.isLoading ? 'message loader' : 'message'}>
 
-      <form onSubmit={formik.handleSubmit}>
-        {isEdit ? (
+      <Modal
+        closeButton
+        aria-labelledby="modal-title"
+        open={isEdit}
+        onClose={closeHandler}
+        width="40vw"
+        css={{ padding: '20px 40px' }}
+      >
+        <form onSubmit={formik.handleSubmit}>
           <>
-            <textarea
-              name="text"
-              rows="3"
-              onChange={formik.handleChange}
-              onBlur={formik.handleBlur}
-              value={formik.values.text}
-              disabled={message.isLoading}
-            />
-            <div style={{ display: 'flex', justifyContent: 'space-around' }}>
-              <div>
-                <p>Incentive (₹)</p>
-                <input
-                  type='number'
-                  name="incentive"
-                  onChange={formik.handleChange}
-                  onBlur={formik.handleBlur}
-                  value={formik.values.number}
-                  style={{ marginBottom: '25px' }}
-                />
-              </div>
-              <div>
-                <p>category</p>
-                <select
-                  style={{ padding: '10px', width: '200px' }}
-                  name="category"
-                  value={formik.values.category}
-                  onChange={formik.handleChange}
-                  onBlur={formik.handleBlur}
-                >
-                  <option value="" label="select">Select</option>
-                  <option label="technical" value="technical">Technical</option>
-                  <option label="management" value="management">Management</option>
-                  <option label="writing" value="writing">Writing</option>
-                  <option label="creativethinking" value="creativethinking">Creative thinking</option>
-                  <option label="supplychain" value="supplychain">Supply chain</option>
-                  <option label="resources" value="resources">Resources</option>
-                  <option label="networking" value="networking">Networking</option>
-                  <option label="other" value="other">Other</option>
-                </select>
-              </div>
-            </div>
-            <input type="hidden" name="id" />
-            {(formik.touched.text && formik.errors.text) || message.error ? (
-              <p className="error">{formik.errors.text || message.error}</p>
-            ) : null}
+            <Modal.Header>
+              <Text id="modal-title" size={18}>
+                <span>Edit</span>
+                <Text b size={18} style={{ color: 'var(--text' }}>
+                  Thread
+                </Text>
+              </Text>
+            </Modal.Header>
+            <Spacer y={1} />
+            <h2 style={{ textAlign: 'left', marginBottom: '10px' }}>Modify your Thread</h2>
+            <Spacer y={1} />
+
+            {
+              page === 1 && (
+                <>
+                  <Spacer y={1} />
+
+                  <Input
+                    clearable
+                    labelPlaceholder='Title'
+                    type='text'
+                    name="title"
+                    bordered
+                    fullWidth
+                    color="primary"
+                    size="lg"
+                    onChange={formik.handleChange}
+                    onBlur={formik.handleBlur}
+                    value={formik.values.title}
+                    style={{ marginBottom: '25px', marginTop: '1rem' }}
+                  />
+                  <Spacer y={1} />
+                  <Textarea
+                    clearable
+                    minRows={20}
+                    maxRows={30}
+                    name="text"
+                    fullWidth
+                    label="Describe the requirement in brief (Supports Markdown)"
+                    onChange={formik.handleChange}
+                    onBlur={formik.handleBlur}
+                    value={formik.values.text}
+                  // disabled={message.isLoading}
+                  />
+                  <input type="hidden" name="id" />
+                  {(formik.touched.text && formik.errors.text) || message.error ? (
+                    <p className="error">{formik.errors.text || message.error}</p>
+                  ) : null}
+                </>
+
+              )}
+
+            {
+              page === 2 && (
+                <div style={{ display: 'flex', justifyContent: 'space-around', flexDirection: 'column', textAlign: 'left' }}>
+
+                  <Spacer y={2} />
+                  <Input
+                    clearable
+                    label='Incentive (₹)'
+                    type='number'
+                    name="incentive"
+                    bordered
+                    fullWidth
+                    color="primary"
+                    size="lg"
+                    onChange={formik.handleChange}
+                    onBlur={formik.handleBlur}
+                    value={formik.values.incentive}
+                    helperText={(new Intl.NumberFormat('en-us', {
+                      style: 'currency',
+                      currency: 'INR',
+                    })).format(formik.values.incentive)}
+                    style={{ marginBottom: '25px', marginTop: '1rem' }}
+                  />
+                  <Spacer y={3} />
+                  <Row justify="start">
+                    <Radio.Group label="Difficulty Of the problem" orientation="horizontal" value={selectedDifficulty} onChange={setSelectedDifficulty}>
+                      <Radio value="easy" description="For Anyone Interested">
+                        Easy
+                      </Radio>
+                      <Radio value="medium" description="People from the field">
+                        Medium
+                      </Radio>
+                      <Radio value="hard" description="Field Experts">
+                        Hard
+                      </Radio>
+                    </Radio.Group>
+                  </Row>
+                  <Spacer y={2} />
+                  <Row justify="start">
+                    <Radio.Group label="Should Answers be public" orientation="horizontal" value={selectPublic} onChange={setselectedPublic}>
+                      <Radio value={false}>
+                        no
+                      </Radio>
+                      <Radio value={true}>
+                        Yes
+                      </Radio>
+                    </Radio.Group>
+                  </Row>
+                  <Spacer y={2} />
+                  <input type="hidden" name="id" />
+                  {(formik.touched.text && formik.errors.text) || message.error ? (
+                    <p className="error">{formik.errors.text || message.error}</p>
+                  ) : null}
+                </div>
+              )
+            }
+
+            {
+              page === 3 && (
+                <>
+                  <Checkbox.Group
+                    label="Fields"
+                    color="secondary"
+                    value={selected}
+                    onChange={setSelected}
+                  >
+                    {Categories.map(e => (
+                      <Checkbox key={e.id} value={e.value} onChange={() => formik.values.category = selected}>{e.name}</Checkbox>
+                    ))}
+                  </Checkbox.Group>
+                </>
+              )
+            }
+
+            <Modal.Footer>
+              <Button auto flat color="error" onPress={closeHandler}>
+                Close
+              </Button>
+              <Button auto flat color="warning" onPress={() => setPage(page - 1)} disabled={page === 1}>
+                Back
+              </Button>
+
+              <Button auto onClick={() => {
+                setPage(page + 1)
+              }
+              } css={{ display: page === 3 ? 'none' : 'block' }}>
+                Next
+              </Button>
+
+
+              <Button type='submit' auto color="success" disabled={message.isLoading} css={{ display: page !== 3 ? 'none' : 'block' }}>
+                Submit
+              </Button>
+
+            </Modal.Footer>
 
           </>
-        ) : (
-          <>
-            <div style={{ display: 'flex', justifyContent: 'space-between' }}>
-              <h2 style={{maxWidth: '70%'}}>{message.text.substring(0, 100)}</h2>
-              <p>{message.incentive ? `₹ ${message.incentive}` : 'undisclosed'}</p>
-            </div>
-            <p>{message.category}</p>
-          </>
-        )}
-        <div className="message-header">
-          <Link to={`/${message.user.username}`}>
-            {/* <img src={message.user.avatar} className="avatar" /> */}
-          </Link>
-          <div>
-            <Link to={`/${message.user.username}`} className="name" style={{ textDecoration: "none", color: "#4040d4" }}>
-              {message.user.name}
-            </Link>
-            <span className="time text-light">{moment(message.createdAt).fromNow()}</span>
-          </div>
+        </form>
+      </Modal>
+
+
+      <>
+        <div style={{ display: 'flex', justifyContent: 'space-between' }}>
+          <h2 style={{ maxWidth: '70%' }}>{message?.title?.substring(0, 100)}</h2>
+          <h3>{message.incentive ? `₹ ${message.incentive}` : 'undisclosed'}</h3>
         </div>
+        <ReactMarkdown>
+          {`${message.text.substring(0, 150)} .....`}
+        </ReactMarkdown>
+        {/* <Spacer y={1}/> */}
         {
-          !auth.me.organisation?.flag && (
+          message?.category?.map(e => (
+            <Badge size='sm' color='primary' variant="flat" style={{ border: '0', cursor: 'pointer', margin: '10px 10px 10px 0px', padding: '8px' }}>{e}</Badge>
+          ))
+        }
+      </>
+
+      <div className="message-header">
+        <Link to={`/${message.user.username}`}>
+          {/* <img src={message.user.avatar} className="avatar" /> */}
+        </Link>
+        <div>
+          <Link to={`/${message.user.username}`} className="name" style={{ textDecoration: "none", color: "#4040d4" }}>
+            {message.user.name}
+          </Link>
+          <span className="time text-light">{moment(message.createdAt).fromNow()}</span>
+        </div>
+      </div>
+      <>
+        <Link to={`/message/${message.id}`}>
+          <div className="btn" style={{ backgroundColor: '#4040d4', color: 'white', cursor: 'pointer' }}>
+            {auth.me.organisation.flag ? 'Preview' : 'Give Solution'}
+          </div>
+        </Link>
+      </>
+      {auth.isAuthenticated && (auth.me.id === message.user.id || auth.me.role === 'ADMIN') && (
+        <>
+          {!isEdit && (
             <>
-              <button onClick={() => { setOpenModal(!openModal) }} type="button" className="btn" style={{ backgroundColor: '#4040d4', color: 'white', cursor: 'pointer' }}>
-                Give Solution
+              <button onClick={handleClickEdit} type="button" className="btn" style={{ backgroundColor: '#4040d4', color: 'white', cursor: 'pointer' }}>
+                Edit
+              </button>
+              <button onClick={(e) => handleDelete(e, message.id)} type="button" className="btn" style={{ backgroundColor: 'red', color: 'white', cursor: 'pointer' }}>
+                Delete
               </button>
             </>
           )
-        }
-        {auth.isAuthenticated && (auth.me.id === message.user.id || auth.me.role === 'ADMIN') && (
-          <>
-            {!isEdit ? (
-              <>
-                <button onClick={() => setOpenSolution(!openSolution)} type="button" className="btn" style={{ backgroundColor: 'green', color: 'white', cursor: 'pointer' }}>
-                  Submissions 
-                </button>
-                <button onClick={handleClickEdit} type="button" className="btn" style={{ backgroundColor: '#4040d4', color: 'white', cursor: 'pointer' }}>
-                  Edit
-                </button>
-                <button onClick={(e) => handleDelete(e, message.id)} type="button" className="btn" style={{ backgroundColor: 'red', color: 'white', cursor: 'pointer' }}>
-                  Delete
-                </button>
-              </>
-            ) : (
-              <>
-                <button type="submit" className="btn" disabled={message.isLoading}>
-                  Submit
-                </button>
-                <button
-                  onClick={() => {
-                    setIsEdit((oldIsEdit) => !oldIsEdit);
-                    clearMessageError(message.id);
-                  }}
-                  type="button"
-                  className="btn"
-                >
-                  Cancel
-                </button>
-
-              </>
-            )}
-          </>
-        )}
-
-        { }
-      </form>
+          }
+        </>
+      )}
 
       {
         openModal && (
@@ -187,7 +297,7 @@ const Message = ({ message, auth, deleteMessage, editMessage, clearMessageError 
 
       {
         openSolution && (
-          <Solutionlist orgSolution={message.id}/>
+          <Solutionlist orgSolution={message.id} />
         )
       }
     </div>
